Add option -x to display non ascii characters in hex format.
The bytes of the fixed properties are displayed with printf and the %c
format directive if the option -g is used.
With the option -x the non ascii characters will be displayed in hex format. Ascii
characters will still be displayed as characters.
The output of non ascii characters caused problems when eltt2 was used in
gitlab-runner tests on raspian with gitlab runner 11.2.0.

Signed-off-by: Juergen Repp <juergen.repp@sit.fraunhofer.de>